requirements.txt
httpx
requests
aiohttp
regex
PyYAML
telebot
uvicorn
fastapi